a_iface_bssap: check bssmap length field
At the moment the length field of the bssmap header is not parsed.
Instead the length is computed out of the known header length and the
number of bytes received. This is prone to error, lets make sure that
extranous data at the end of a message is ignored by parsing the bssmap
length correctly.

Change-Id: I3b89dd5a66ec83b03860b58b6b8eb58007f433a4
Related: OS#3806