make vty-transcrip-test: use $VTY_TEST var like osmo-bsc
Adopt the same way to run manual vty transcript tests as in
osmo-bsc/test/Makefile.am.

There are different ways to select a specific test to run in osmo-bsc and here
in osmo-msc. The osmo-bsc way is more convenient when building outside the src
tree, because it does not need the full absolute path of the test file.

Change-Id: If1e2abfa321a5e9fb60358d1f0e4e448b33184af