After RX of an SMPP Submit, send the SMS we just received.
Previously, after we have stored the new SMS in the db, we then
signalled the queue to run, but this is not good enough.

The queue may not pick up this new SMS on the next run,
if there are sufficient older messages in the queue. So let's
specifically send the message that was just submitted.

Change-Id: I9af51ef0d9c2e6c5acc5128efd6195df881b680c